+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Released]
 
+## [3.1.5] - 2021-04-30
+
+### Added
+
+- see [pull request](https://github.com/hiranthaR/Json-to-Dart-Model/pull/34) from [Arnas](https://github.com/iamarnas)
+   - Improved equality operator.
+   - Added annotations support.
+
+   It is possible to mark `JSON` values as default or required by adding to your `JSON` key `d@` or `r@`
+
 ## [3.1.3] - 2021-04-23
 
 ### Fixed